refactor: moved some code around
