+++ .withlist/server_ed25519.pub
@@ -1,0 +1,1 @@
+ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAINp6BT/a5T/BY6p3NEU26yTJ8OFkRBhkINxfDfrjdl3g carlos@darkstar
