+++ client_local.go
@@ -119,7 +119,7 @@ func (c *localClient) Connect(e *Endpoint) (tea.ExecCommand, error) {
 		return nil, fmt.Errorf("failed to setup a authentication method: %w", err)
 	}
 
-	session, client, cls, err := createSession(&ssh.ClientConfig{
+	session, client, closers, err := createSession(&ssh.ClientConfig{
 		User:            firstNonEmpty(e.User, user.Username),
 		Auth:            methods,
 		HostKeyCallback: hostKeyCallback(e, filepath.Join(user.HomeDir, ".ssh/known_hosts")),
@@ -132,6 +132,6 @@ func (c *localClient) Connect(e *Endpoint) (tea.ExecCommand, error) {
 		endpoint: e,
 		session:  session,
 		client:   client,
-		closers:  cls,
+		closers:  closers,
 	}, err
 }
