+++ multiplex/reader_test.go
@@ -1,4 +1,4 @@
-package wishlist
+package multiplex
 
 import (
 	"bytes"
@@ -19,7 +19,7 @@ func TestMultiplex(t *testing.T) {
 
 		done := make(chan bool, 1)
 		t.Cleanup(func() { done <- true })
-		r1, r2 := multiplex(&b, done)
+		r1, r2 := Reader(&b, done)
 
 		b1, err := io.ReadAll(r1)
 		require.NoError(t, err)
@@ -34,7 +34,7 @@ func TestMultiplex(t *testing.T) {
 		r := iotest.ErrReader(io.ErrClosedPipe)
 		done := make(chan bool, 1)
 		t.Cleanup(func() { done <- true })
-		r1, r2 := multiplex(r, done)
+		r1, r2 := Reader(r, done)
 
 		b1, err := io.ReadAll(r1)
 		require.NoError(t, err)
